+++ components/drawer/__tests__/DrawerEvent.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Drawer from '..';
-import { render, fireEvent } from '../../../tests/utils';
+
+import { render, fireEvent, screen } from '../../../tests/utils';
 
 describe('Drawer', () => {
   const getDrawer = props => (
@@ -44,5 +45,32 @@ describe('Drawer', () => {
 
     fireEvent.click(container.querySelector('.ant-drawer-mask'));
     expect(onClose).not.toHaveBeenCalled();
+  });
+
+  it('dom should be removed after close when destroyOnClose is true', () => {
+    const { rerender, container } = render(getDrawer({ destroyOnClose: true }));
+    rerender(getDrawer({ destroyOnClose: true, visible: false }));
+    const ev = new Event('transitionend', { bubbles: true });
+    ev.propertyName = 'transform';
+    fireEvent(container.querySelector('.ant-drawer-content-wrapper'), ev);
+    expect(screen.queryByText(/Here is content of Drawer/)).toBeFalsy();
+  });
+  it('dom should be existed after close when destroyOnClose is false', () => {
+    const { rerender, container } = render(getDrawer());
+    expect(screen.queryByText(/Here is content of Drawer/)).toBeTruthy();
+    rerender(getDrawer({ visible: false }));
+    const ev = new Event('transitionend', { bubbles: true });
+    ev.propertyName = 'transform';
+    fireEvent(container.querySelector('.ant-drawer-content-wrapper'), ev);
+    expect(screen.queryByText(/Here is content of Drawer/)).toBeTruthy();
+  });
+  it('test afterVisibleChange', async () => {
+    const afterVisibleChange = jest.fn();
+    const { rerender, container } = render(getDrawer({ afterVisibleChange, visible: true }));
+    rerender(getDrawer({ afterVisibleChange, visible: false }));
+    const ev = new Event('transitionend', { bubbles: true });
+    ev.propertyName = 'transform';
+    fireEvent(container.querySelector('.ant-drawer-content-wrapper'), ev);
+    expect(afterVisibleChange).toBeCalledTimes(1);
   });
 });
